+++ changelog.md
@@ -4,6 +4,7 @@
 
 - added: comments to user saved screen
 - added: remember sorting preference
+- added: drawer everywhere
 - fixed: gallery flashes previous image
 
 ## Build 40